* Update to 0.7.7 * Move from minium build to maximum (change WITH_ to WITHOUT_) by default * Allow to build with local-cache and virus-scan features * make startup script more smarter
PR:		50674
Submitted by:	Sergey Matveychuk <sem@ciam.ru>

+++ ftp/frox/files/frox.sh
@@ -8,12 +8,12 @@ fi
 case "$1" in
 start)
 	[ -x ${PREFIX}/sbin/frox -a -f ${PREFIX}/etc/frox.conf ]
-	        grep -q '^ *# *FromInetd *yes' ${PREFIX}/etc/frox.conf && \
+	        ! grep -iq '^[[:space:]]*frominetd[[:space:]]*yes' ${PREFIX}/etc/frox.conf && \
 		${PREFIX}/sbin/frox && \
 		echo -n ' frox'
 	;;
 stop)
-	kill `cat /var/run/frox.pid` && echo -n ' frox'
+	[ -f /var/run/frox.pid ] && kill `cat /var/run/frox.pid` && echo -n ' frox'
 	;;
 *)
 	echo "Usage: `basename $0` {start|stop}" >&2
